+++ src/pages/Menu.tsx
@@ -24,7 +24,7 @@ import Details from './Details';
     const path = [
         {name:'Home', url: '/it35-lab/app/home', icon: homeOutline},
         {name:'About', url: '/it35-lab/app/about', icon: rocketOutline},
-        {name:'Details', url: '/it35-lab/app/about', icon: homeOutline},
+        
     ]
 
     return (
